+++ demo-projects/blog/initialData.js
@@ -1,0 +1,18 @@
+module.exports = {
+  User: [
+    {
+      name: 'Administrator',
+      email: 'admin@keystone.project',
+      isAdmin: true,
+      dob: '1990-01-01',
+      password: 'password',
+    },
+    {
+      name: 'Demo User',
+      email: 'a@demo.user',
+      isAdmin: false,
+      dob: '1995-06-09',
+      password: 'password',
+    },
+  ],
+};
