+++ demo-projects/blog/schema.js
@@ -1,0 +1,148 @@
+const { LocalFileAdapter } = require('@keystone-alpha/file-adapters');
+const {
+  File,
+  Text,
+  Relationship,
+  Select,
+  Password,
+  Checkbox,
+  CalendarDay,
+  DateTime,
+} = require('@keystone-alpha/fields');
+
+const { staticRoute, staticPath } = require('./config');
+const getYear = require('date-fns/get_year');
+
+const fileAdapter = new LocalFileAdapter({
+  directory: `${staticPath}/uploads`,
+  route: `${staticRoute}/uploads`,
+});
+
+const avatarFileAdapter = new LocalFileAdapter({
+  directory: `${staticPath}/avatars`,
+  route: `${staticRoute}/avatars`,
+});
+
+exports.createPublicAPI = ({ keystone }) => {
+  const read = true;
+  const create = true;
+  return keystone.createGraphQLMiddleware({
+    lists: {
+      User: { fields: { name: { read }, avatar: { read } } },
+      Post: { access: { read: { status: 'published' } } },
+      PostCategory: { access: { read } },
+      Comment: { access: read, create },
+    },
+  });
+};
+
+exports.createLists = ({ keystone }) => {
+  keystone.createList('User', {
+    fields: {
+      name: {
+        type: Text,
+      },
+      email: {
+        type: Text,
+        isUnique: true,
+      },
+      dob: {
+        type: CalendarDay,
+        format: 'Do MMMM YYYY',
+        yearRangeFrom: 1901,
+        yearRangeTo: getYear(new Date()),
+      },
+      password: {
+        type: Password,
+      },
+      isAdmin: {
+        type: Checkbox,
+      },
+      avatar: {
+        type: File,
+        adapter: avatarFileAdapter,
+      },
+    },
+    labelResolver: item => `${item.name}`,
+  });
+
+  keystone.createList('Post', {
+    fields: {
+      title: {
+        type: Text,
+      },
+      author: {
+        type: Relationship,
+        ref: 'User',
+      },
+      categories: {
+        type: Relationship,
+        ref: 'PostCategory',
+        many: true,
+      },
+      status: {
+        type: Select,
+        defaultValue: 'draft',
+        options: [
+          {
+            label: 'Draft',
+            value: 'draft',
+          },
+          {
+            label: 'Published',
+            value: 'published',
+          },
+        ],
+      },
+      body: {
+        type: Text,
+        isMultiline: true,
+      },
+      posted: {
+        type: DateTime,
+      },
+      image: {
+        type: File,
+        adapter: fileAdapter,
+      },
+    },
+    adminConfig: {
+      defaultPageSize: 20,
+      defaultColumns: 'title, status',
+      defaultSort: 'title',
+    },
+    labelResolver: item => item.title,
+  });
+
+  keystone.createList('PostCategory', {
+    fields: {
+      name: {
+        type: Text,
+      },
+      slug: {
+        type: Text,
+      },
+    },
+  });
+
+  keystone.createList('Comment', {
+    fields: {
+      body: {
+        type: Text,
+        isMultiline: true,
+      },
+      originalPost: {
+        type: Relationship,
+        ref: 'Post',
+      },
+      author: {
+        type: Relationship,
+        ref: 'User',
+      },
+      posted: {
+        type: DateTime,
+      },
+    },
+    labelResolver: item => item.body,
+  });
+};
